apps/crew/requirements.txt
crewai[tools]==0.80.0
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.28.1
anthropic==0.40.0
openai==1.58.1
python-dotenv==1.0.1
